+++ dpypelines/pipeline/shared/transforms/sdmx/v20.py
@@ -1,9 +1,9 @@
 from pathlib import Path
 
-from dpypelines.pipeline.shared.transforms.sdmx.compact.v20.prototype.v1 import (
-    generate_versions_metadata,
-    xmlToCsvSDMX2_0,
-)
+from dpypelines.pipeline.shared.transforms.sdmx.generate_versions_metadata import generate_versions_metadata
+
+from dpypelines.pipeline.shared.transforms.sdmx.compact.v20.prototype.v1 import xmlToCsvSDMX2_0
+
 
 
 def sdmx_compact_2_0_prototype_1(input_file: Path):
