+++ dpypelines/pipeline/shared/transforms/sdmx/v21.py
@@ -1,9 +1,8 @@
 from pathlib import Path
 
-from dpypelines.pipeline.shared.transforms.sdmx.generic.v21.prototype.v1 import (
-    generate_versions_metadata,
-    xmlToCsvSDMX2_1,
-)
+from dpypelines.pipeline.shared.transforms.sdmx.generate_versions_metadata import generate_versions_metadata
+
+from dpypelines.pipeline.shared.transforms.sdmx.generic.v21.prototype.v1 import xmlToCsvSDMX2_1
 
 
 def sdmx_generic_2_1_prototype_1(input_file: Path):
